server/requirements.txt
diffusers==0.26.3
transformers==4.38.2
accelerate==0.27.2
huggingface-hub==0.22.2
torch
pillow
fastapi
uvicorn